requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
colorama==0.4.4
furl==2.1.3
lxml==4.9.2
pyTelegramBotAPI==4.10.0
PyYAML==6.0
requests==2.27.1